renew

+++ torch/csrc/jit/passes/common_subexpression_elimination.cpp
@@ -1,19 +1,15 @@
-#include "torch/csrc/jit/ir.h"
+#include <torch/csrc/jit/passes/common_subexpression_elimination.h>
 
-#include <algorithm>
+#include <torch/csrc/jit/ir.h>
+#include <torch/csrc/jit/node_hashing.h>
+#include <torch/csrc/jit/passes/alias_analysis.h>
+
 #include <unordered_map>
-
-#include "torch/csrc/jit/assertions.h"
-#include "torch/csrc/jit/interned_strings.h"
-#include "torch/csrc/jit/passes/alias_analysis.h"
-#include "torch/csrc/jit/passes/common_subexpression_elimination.h"
-#include "torch/csrc/jit/node_hashing.h"
-#include "torch/csrc/utils/functional.h"
-#include "torch/csrc/utils/hash.h"
 
 namespace torch {
 namespace jit {
 namespace {
+
 // The function implements common subexpression elimination.
 // Since the nodes are visited in topological order, one pass is enough.
 void EliminateCommonSubexpression(
@@ -21,9 +17,9 @@ void EliminateCommonSubexpression(
     const AliasDb& aliasDb,
     std::function<Node*(Node*)> parent_lookup_fn) {
   std::unordered_set<Node*, HashNode, EqualNode> subexprs;
-  for (auto it = block->nodes().begin(); it != block->nodes().end(); ++ it) {
+  for (auto it = block->nodes().begin(); it != block->nodes().end(); ++it) {
     auto node = *it;
-    if (node->kind() == prim::PythonOp || node->kind() == prim::Print ||
+    if (node->hasSideEffects() || node->isNondeterministic() ||
         aliasDb.hasWriters(node)) {
       // Do NOT have enough information to do CSE on these nodes.
       continue;
@@ -47,7 +43,15 @@ void EliminateCommonSubexpression(
 
     // Check for CSE opportunities in the parent block.
     auto parent_lookup = parent_lookup_fn(node);
+    auto g_out = node->owningGraph()->outputs();
     if (parent_lookup) {
+      // since the graph outputs may be mutated after they are returned,
+      // don't introduce new aliasing among graph outputs
+      if (aliasDb.mayContainAlias(node->outputs(), g_out) &&
+          aliasDb.mayContainAlias(parent_lookup->outputs(), g_out)) {
+        continue;
+      }
+
       node->replaceAllUsesWith(parent_lookup);
       it.destroyCurrent();
       continue;
@@ -58,16 +62,24 @@ void EliminateCommonSubexpression(
     if (!subit.second) {
       // Subexpression exists, replace the uses of node, and destroy it.
       auto existing = *subit.first;
+
+      // don't introduce new aliasing among graph outputs
+      if (aliasDb.mayContainAlias(
+              node->outputs(), node->owningGraph()->outputs()) &&
+          aliasDb.mayContainAlias(existing->outputs(), g_out)) {
+        continue;
+      }
+
       node->replaceAllUsesWith(existing);
       // Destroy the node.
       it.destroyCurrent();
     }
   }
 }
-}
+} // namespace
 
 void EliminateCommonSubexpression(std::shared_ptr<Graph>& graph) {
-  const auto aliasDb = AliasAnalysis(graph);
+  AliasDb aliasDb(graph);
   EliminateCommonSubexpression(
       graph->block(), aliasDb, [](Node*) { return nullptr; });
 }
